+++ packages/components/psammead-brand/src/index.jsx
@@ -4,7 +4,7 @@ import { string, number, node, shape, bool } from 'prop-types';
 import { C_POSTBOX, C_WHITE } from '@bbc/psammead-styles/colours';
 import VisuallyHiddenText from '@bbc/psammead-visually-hidden-text';
 import {
-  GEL_GROUP_3_SCREEN_WIDTH_MIN,
+  GEL_GROUP_2_SCREEN_WIDTH_MIN,
   GEL_GROUP_5_SCREEN_WIDTH_MIN,
 } from '@bbc/gel-foundations/breakpoints';
 import {
@@ -14,10 +14,10 @@ import {
   GEL_SPACING_TRPL,
 } from '@bbc/gel-foundations/spacings';
 
-const SVG_TOP_OFFSET_ABOVE_600PX = '1.75rem'; // 28px
-const SVG_BOTTOM_OFFSET_BELOW_600PX = '0.75rem'; // 12px
-const PADDING_AROUND_SVG_ABOVE_600PX = 56;
-const PADDING_AROUND_SVG_BELOW_600PX = 32;
+const SVG_TOP_OFFSET_ABOVE_400PX = '1.75rem'; // 28px
+const SVG_BOTTOM_OFFSET_BELOW_400PX = '0.75rem'; // 12px
+const PADDING_AROUND_SVG_ABOVE_400PX = 56;
+const PADDING_AROUND_SVG_BELOW_400PX = 32;
 
 const conditionallyRenderHeight = (svgHeight, padding) =>
   svgHeight ? `height: ${(svgHeight + padding) / 16}rem` : '';
@@ -32,13 +32,13 @@ const SvgWrapper = styled.div`
 const Banner = styled.div`
   background-color: ${C_POSTBOX};
   ${({ svgHeight }) =>
-    conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_BELOW_600PX)};
+    conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_BELOW_400PX)};
   width: 100%;
   padding: 0 ${GEL_SPACING};
 
-  @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) {
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) {
     ${({ svgHeight }) =>
-      conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_ABOVE_600PX)};
+      conditionallyRenderHeight(svgHeight, PADDING_AROUND_SVG_ABOVE_400PX)};
     padding: 0 ${GEL_SPACING_DBL};
   }
   border-top: ${({ borderTop }) => borderTop && TRANSPARENT_BORDER};
@@ -56,16 +56,19 @@ const StyledLink = styled.a`
   ${({ maxWidth, minWidth }) => brandWidth(minWidth, maxWidth)}
 `;
 
+// `currentColor` has been used to address high contrast mode in Firefox.
 const BrandSvg = styled.svg`
   box-sizing: content-box;
-  fill: ${C_WHITE};
+  color: ${C_WHITE};
+  fill: currentColor;
   padding-top: ${GEL_SPACING_DBL};
-  padding-bottom: ${SVG_BOTTOM_OFFSET_BELOW_600PX};
-
+  padding-bottom: ${SVG_BOTTOM_OFFSET_BELOW_400PX};
+  height: ${props => props.height / 16}rem;
+  
   ${({ maxWidth, minWidth }) => brandWidth(minWidth, maxWidth)}
 
-  @media (min-width: ${GEL_GROUP_3_SCREEN_WIDTH_MIN}) {
-    padding-top: ${SVG_TOP_OFFSET_ABOVE_600PX};
+  @media (min-width: ${GEL_GROUP_2_SCREEN_WIDTH_MIN}) {
+    padding-top: ${SVG_TOP_OFFSET_ABOVE_400PX};
     padding-bottom: ${GEL_SPACING_TRPL};
   }
 
@@ -85,11 +88,12 @@ const BrandSvg = styled.svg`
 
 const LocalisedBrandName = ({ product, serviceLocalisedName }) =>
   serviceLocalisedName ? (
-    <Fragment>
+    // eslint-disable-next-line jsx-a11y/aria-role
+    <VisuallyHiddenText role="text">
       <span lang="en-GB">{product}</span>, {serviceLocalisedName}
-    </Fragment>
+    </VisuallyHiddenText>
   ) : (
-    product
+    <VisuallyHiddenText>{product}</VisuallyHiddenText>
   );
 
 LocalisedBrandName.propTypes = {
@@ -124,12 +128,10 @@ const StyledBrand = ({
         >
           {svg.group}
         </BrandSvg>
-        <VisuallyHiddenText>
-          <LocalisedBrandName
-            product={product}
-            serviceLocalisedName={serviceLocalisedName}
-          />
-        </VisuallyHiddenText>
+        <LocalisedBrandName
+          product={product}
+          serviceLocalisedName={serviceLocalisedName}
+        />
       </Fragment>
     )}
   </Fragment>