+++ packages/utilities/psammead-storybook-helpers/src/index.test.jsx
@@ -1,5 +1,6 @@
 import { select, text } from '@storybook/addon-knobs';
 import { testUtilityPackages } from '@bbc/psammead-test-helpers';
+import textVariants from './text-variants';
 import * as underTest from '.';
 
 jest.mock('@storybook/addon-knobs');
@@ -53,45 +54,66 @@ describe('Psammead storybook helpers', () => {
     });
 
     it('always calls the render function with a script and direction', () => {
-      const english = {
-        text: 'Could a computer ever create better art than a human?',
-        script: 'latin',
-        dir: 'ltr',
-      };
-      select.mockReturnValueOnce(english);
+      select.mockReturnValueOnce('news');
 
       underTest.inputProvider([], renderFn)();
 
       expect(renderFn).toHaveBeenCalledTimes(1);
-      expect(renderFn).toHaveBeenCalledWith([], 'LATIN SCRIPT OBJECT', 'ltr');
+      expect(renderFn).toHaveBeenCalledWith({
+        dir: 'ltr',
+        slotTexts: [],
+        script: 'LATIN SCRIPT OBJECT',
+        service: 'news',
+      });
       expect(select).toHaveBeenCalledTimes(1);
+      // Allows user to select all availible services
+      expect(select.mock.calls[0][1].length).toEqual(
+        Object.keys(textVariants).length,
+      );
+      expect(text).toHaveBeenCalledTimes(0);
+    });
+
+    it('filters services availible based on array provided', () => {
+      select.mockReturnValueOnce('news');
+
+      underTest.inputProvider([], renderFn, [
+        'mundo',
+        'pidgin',
+        'yoruba',
+        'foobar',
+      ])();
+
+      expect(renderFn).toHaveBeenCalledTimes(1);
+      expect(renderFn).toHaveBeenCalledWith({
+        dir: 'ltr',
+        slotTexts: [],
+        script: 'LATIN SCRIPT OBJECT',
+        service: 'news',
+      });
+      expect(select).toHaveBeenCalledTimes(1);
+      expect(select.mock.calls[0][1]).toEqual(['mundo', 'pidgin', 'yoruba']);
       expect(text).toHaveBeenCalledTimes(0);
     });
 
     it('handles scenario where config is null', () => {
-      const english = {
-        text: 'Could a computer ever create better art than a human?',
-        script: 'latin',
-        dir: 'ltr',
-      };
-      select.mockReturnValueOnce(english);
+      select.mockReturnValueOnce('news');
 
       underTest.inputProvider(null, renderFn)();
 
       expect(renderFn).toHaveBeenCalledTimes(1);
-      expect(renderFn).toHaveBeenCalledWith([], 'LATIN SCRIPT OBJECT', 'ltr');
+      expect(renderFn).toHaveBeenCalledWith({
+        dir: 'ltr',
+        slotTexts: [],
+        script: 'LATIN SCRIPT OBJECT',
+        service: 'news',
+      });
       expect(select).toHaveBeenCalledTimes(1);
       expect(text).toHaveBeenCalledTimes(0);
     });
 
-    describe('calls the render function with slot default text when displaying english', () => {
+    describe('calls the render function with slot default text when displaying news', () => {
       it('for a single slot', () => {
-        const english = {
-          text: 'Could a computer ever create better art than a human?',
-          script: 'latin',
-          dir: 'ltr',
-        };
-        select.mockReturnValueOnce(english);
+        select.mockReturnValueOnce('news');
         text.mockImplementation((_, displayText) => displayText);
 
         underTest.inputProvider(
@@ -100,23 +122,19 @@ describe('Psammead storybook helpers', () => {
         )();
 
         expect(renderFn).toHaveBeenCalledTimes(1);
-        expect(renderFn).toHaveBeenCalledWith(
-          ['Sole input'],
-          'LATIN SCRIPT OBJECT',
-          'ltr',
-        );
+        expect(renderFn).toHaveBeenCalledWith({
+          slotTexts: ['Sole input'],
+          script: 'LATIN SCRIPT OBJECT',
+          dir: 'ltr',
+          service: 'news',
+        });
 
         expect(select).toHaveBeenCalledTimes(1);
         expect(text).toHaveBeenCalledTimes(1);
       });
 
       it('for multiple slots', () => {
-        const english = {
-          text: 'Could a computer ever create better art than a human?',
-          script: 'latin',
-          dir: 'ltr',
-        };
-        select.mockReturnValueOnce(english);
+        select.mockReturnValueOnce('news');
         text.mockImplementation((_, displayText) => displayText);
 
         underTest.inputProvider(
@@ -128,11 +146,12 @@ describe('Psammead storybook helpers', () => {
         )();
 
         expect(renderFn).toHaveBeenCalledTimes(1);
-        expect(renderFn).toHaveBeenCalledWith(
-          ['First input', 'Second input'],
-          'LATIN SCRIPT OBJECT',
-          'ltr',
-        );
+        expect(renderFn).toHaveBeenCalledWith({
+          slotTexts: ['First input', 'Second input'],
+          script: 'LATIN SCRIPT OBJECT',
+          dir: 'ltr',
+          service: 'news',
+        });
 
         expect(select).toHaveBeenCalledTimes(1);
         expect(text).toHaveBeenCalledTimes(2);
@@ -140,35 +159,25 @@ describe('Psammead storybook helpers', () => {
     });
 
     it('displays english snippet when no slot default text provided', () => {
-      const english = {
-        text: 'Could a computer ever create better art than a human?',
-        script: 'latin',
-        dir: 'ltr',
-      };
-      select.mockReturnValueOnce(english);
+      select.mockReturnValueOnce('news');
       text.mockImplementation((_, displayText) => displayText);
 
       underTest.inputProvider([{ name: 'first' }], renderFn)();
 
       expect(renderFn).toHaveBeenCalledTimes(1);
-      expect(renderFn).toHaveBeenCalledWith(
-        ['Could a computer ever create better art than a human?'],
-        'LATIN SCRIPT OBJECT',
-        'ltr',
-      );
+      expect(renderFn).toHaveBeenCalledWith({
+        slotTexts: ['Could a computer ever create better art than a human?'],
+        script: 'LATIN SCRIPT OBJECT',
+        dir: 'ltr',
+        service: 'news',
+      });
 
       expect(select).toHaveBeenCalledTimes(1);
       expect(text).toHaveBeenCalledTimes(1);
     });
 
-    it('defaults to language text for non-english languages', () => {
-      const russian = {
-        text:
-          'Мнение Назарбаева будет иметь приоритетное значение при принятии важных для страны решений, сказал Токаев',
-        script: 'cyrillic',
-        dir: 'ltr',
-      };
-      select.mockReturnValueOnce(russian);
+    it('defaults to service text for non-news services', () => {
+      select.mockReturnValueOnce('russian');
       text.mockImplementation((_, displayText) => displayText);
 
       underTest.inputProvider(
@@ -177,66 +186,59 @@ describe('Psammead storybook helpers', () => {
       )();
 
       expect(renderFn).toHaveBeenCalledTimes(1);
-      expect(renderFn).toHaveBeenCalledWith(
-        [
+      expect(renderFn).toHaveBeenCalledWith({
+        slotTexts: [
           'Мнение Назарбаева будет иметь приоритетное значение при принятии важных для страны решений, сказал Токаев',
         ],
-        'CYRILLIC SCRIPT OBJECT',
-        'ltr',
-      );
+        script: 'CYRILLIC SCRIPT OBJECT',
+        dir: 'ltr',
+        service: 'russian',
+      });
     });
 
     describe('text direction', () => {
       it('defaults to ltr when not specified', () => {
-        const russian = {
-          script: 'cyrillic',
-        };
-        select.mockReturnValueOnce(russian);
+        select.mockReturnValueOnce('russian');
         text.mockImplementation((textKnobName, displayText) => displayText);
 
         underTest.inputProvider([], renderFn)();
 
         expect(renderFn).toHaveBeenCalledTimes(1);
-        expect(renderFn).toHaveBeenCalledWith(
-          [],
-          'CYRILLIC SCRIPT OBJECT',
-          'ltr',
-        );
+        expect(renderFn).toHaveBeenCalledWith({
+          slotTexts: [],
+          script: 'CYRILLIC SCRIPT OBJECT',
+          dir: 'ltr',
+          service: 'russian',
+        });
       });
 
       it('returns value when specified', () => {
-        const arabic = {
-          script: 'arabic',
-          dir: 'rtl',
-        };
-        select.mockReturnValueOnce(arabic);
+        select.mockReturnValueOnce('arabic');
         text.mockImplementation((_, displayText) => displayText);
 
         underTest.inputProvider([], renderFn)();
 
         expect(renderFn).toHaveBeenCalledTimes(1);
-        expect(renderFn).toHaveBeenCalledWith(
-          [],
-          'ARABIC SCRIPT OBJECT',
-          'rtl',
-        );
+        expect(renderFn).toHaveBeenCalledWith({
+          slotTexts: [],
+          script: 'ARABIC SCRIPT OBJECT',
+          dir: 'rtl',
+          service: 'arabic',
+        });
       });
     });
 
     describe('dirDecorator', () => {
       it('calls the story function with dir and script', () => {
         const storyFn = jest.fn();
-        const english = {
-          script: 'latin',
-          dir: 'ltr',
-        };
-        select.mockReturnValueOnce(english);
+        select.mockReturnValueOnce('news');
 
         underTest.dirDecorator(storyFn);
 
         expect(storyFn).toHaveBeenCalledWith({
           dir: 'ltr',
           script: 'LATIN SCRIPT OBJECT',
+          service: 'news',
         });
       });
     });