+++ packages/components/psammead-paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 2.0.2   | [PR#892](https://github.com/bbc/psammead/pull/892) Bump dependencies |
+| 2.0.1   | [PR#1019](https://github.com/bbc/psammead/pull/1019) Bump psammead-inline-link to revert styling changes |
+| 2.0.0   | [PR#941](https://github.com/bbc/psammead/pull/941) Apply font based on service prop |
+| 1.0.9   | [PR#783](https://github.com/bbc/psammead/pull/783) Update to latest psammead-test-helpers. Update snapshots. |
+| 1.0.8   | [PR#769](https://github.com/bbc/psammead/pull/769) Fix stories not appearing in storybook when using `install:packages:link` |
+| 1.0.7   | [PR#746](https://github.com/bbc/psammead/pull/746) Use `@bbc/psammead-inline-link@1.1.1` and add a 'with inline link' story |
+| 1.0.6   | [PR#713](https://github.com/bbc/psammead/pull/713) Update `styled-components` to 4.3.2 |
+| 1.0.5   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
 | 1.0.4   | [PR#498](https://github.com/bbc/psammead/pull/498) Update stories to use new input provider |
 | 1.0.3   | [PR#459](https://github.com/bbc/psammead/pull/459) Add custom propTypes shape |
 | 1.0.2   | [PR#424](https://github.com/bbc/psammead/pull/424) Add Snyk badge to readme |