+++ packages/components/psammead-section-label/CHANGELOG.md
@@ -3,6 +3,12 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 1.0.5   | [PR#892](https://github.com/bbc/psammead/pull/892) Bump dependencies |
+| 1.0.4 | [PR#783](https://github.com/bbc/psammead/pull/783) Update to latest psammead-test-helpers. Update snapshots. |
+| 1.0.3   | [PR#769](https://github.com/bbc/psammead/pull/769) Fix stories not appearing in storybook when using `install:packages:link` |
+| 1.0.2   | [PR#713](https://github.com/bbc/psammead/pull/713) Update `styled-components` to 4.3.2 |
+| 1.0.1   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0 | [PR#679](https://github.com/bbc/psammead/pull/679) Bump version number |
 | 0.1.1 | [PR#610](https://github.com/bbc/psammead/pull/610) Add visuallyHidden prop. |
 | 0.1.0 | [PR#616](https://github.com/BBC/psammead/pull/616) Remove alpha tag |
 | 0.1.0-alpha.1 | [PR#611](https://github.com/BBC/psammead/pull/611) Use border top for the horizontal bar |