+++ packages/components/psammead-story-promo/CHANGELOG.md
@@ -3,11 +3,20 @@
 <!-- prettier-ignore -->
 | Version | Description |
 | ------- | ----------- |
+| 1.0.1  | [PR#892](https://github.com/bbc/psammead/pull/892) Bump dependencies |
+| 1.0.0 | [PR#937](https://github.com/bbc/psammead/pull/937) Remove alpha tag, swarm done |
+| 1.0.0-alpha.6   | [PR#893](https://github.com/bbc/psammead/pull/893) Fix knobs for story promo stories |
+| 1.0.0-alpha.5   | [PR#769](https://github.com/bbc/psammead/pull/769) Fix stories not appearing in storybook when using `install:packages:link` |
+| 1.0.0-alpha.4 | [PR#706](https://github.com/bbc/psammead/pull/706) Fix amp images not showing in IE11  |
+| 1.0.0-alpha.3   | [PR#713](https://github.com/bbc/psammead/pull/713) Update `styled-components` to 4.3.2 |
+| 1.0.0-alpha.2 | [PR#694](https://github.com/bbc/psammead/pull/694) Fix floating media indicator and promo link bugs  |
+| 1.0.0-alpha.1 | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0-alpha.0 | [PR#679](https://github.com/bbc/psammead/pull/679) Bump version number |
 | 0.2.0-alpha.9 | [PR#663](https://github.com/bbc/psammead/pull/663) Add assertion tests |
 | 0.2.0-alpha.8 | [PR#578](https://github.com/bbc/psammead/pull/578) Add Top Story promo design |
-| 0.2.0-alpha.7 | [PR#628](https://github.com/bbc/psammead/pull/628) Update Story Promo Changelog title |
+| 0.2.0-alpha.7 | [PR#628](https://github.com/bbc/psammead/pull/628) Update Story  Promo Changelog title |
 | 0.2.0-alpha.6 | [PR#588](https://github.com/bbc/psammead/pull/588) Update Story promo headline to use Pica under 600px. |
-| 0.2.0-alpha.5 | [PR#522](https://github.com/bbc/psammead/pull/522) Add MediaIndicator. |
+| 0.2.0-alpha.5 | [PR#522](https://github.com/bbc/psammead/pull/522) AddMediaIndicator. |
 | 0.2.0-alpha.4 | [PR#546](https://github.com/bbc/psammead/pull/546) Replace colours and hide summary under 600px width screen. |
 | 0.2.0-alpha.3 | [PR#534](https://github.com/BBC-News/psammead/pull/534) Remove Timestamp padding. |
 | 0.2.0-alpha.2 | [PR#535](https://github.com/bbc/psammead/pull/535) Fix issues with knobs on storybook. |