+++ packages/components/psammead-timestamp/CHANGELOG.md
@@ -3,6 +3,12 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 1.0.5   | [PR#892](https://github.com/bbc/psammead/pull/892) Bump dependencies |
+| 1.0.4 | [PR#783](https://github.com/bbc/psammead/pull/783) Update to latest psammead-test-helpers. Update snapshots. |
+| 1.0.3   | [PR#769](https://github.com/bbc/psammead/pull/769) Fix stories not appearing in storybook when using `install:packages:link` |
+| 1.0.2   | [PR#713](https://github.com/bbc/psammead/pull/713) Update `styled-components` to 4.3.2 |
+| 1.0.1   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
+| 1.0.0   | [PR#679](https://github.com/bbc/psammead/pull/679) Bump version number |
 | 0.3.4   | [PR#661](https://github.com/bbc/psammead/pull/661) Update @bbc/gel-foundations to version 3.0.0 |
 | 0.3.3   | [PR#587](https://github.com/bbc/psammead/pull/587) Update colour to C_METAL |
 | 0.3.2   | [PR#515](https://github.com/bbc/psammead/pull/515) Update story to use dirDecorator |