+++ packages/utilities/gel-foundations/CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 3.0.2 | [PR#783](https://github.com/bbc/psammead/pull/783) Update to latest psammead-test-helpers. Update snapshots. |
+| 3.0.1 | [PR#701](https://github.com/bbc/psammead/pull/701) Update README `rem` doc link |
 | 3.0.0 | [PR#586](https://github.com/bbc/psammead/pull/586) Update GEL Types values |
 | 2.0.0 | [PR#517](https://github.com/bbc/psammead/pull/517) Remove unused GEL Types |
 | 1.2.0 | [PR#431](https://github.com/bbc/psammead/pull/431) Add custom scripts propTypes |