+++ packages/utilities/psammead-storybook-helpers/CHANGELOG.md
@@ -3,6 +3,10 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 3.1.1 | [PR#783](https://github.com/bbc/psammead/pull/783) Update to latest psammead-test-helpers. Update snapshots. |
+| 3.1.0   | [PR#758](https://github.com/bbc/psammead/pull/758) Add ability to limit set of services |
+| 3.0.0   | [PR#674](https://github.com/bbc/psammead/pull/674) Update inputProvider to pass an object with script, service, dir and slotTexts |
+| 2.1.2   | [PR#677](https://github.com/bbc/psammead/pull/677) Use `@bbc/gel-foundations@3.0.0` |
 | 2.1.1   | [PR#512](https://github.com/bbc/psammead/pull/512) Pass script and dir to storybook function |
 | 2.1.0   | [PR#496](https://github.com/bbc/psammead/pull/496) Add dir storybook decorator |
 | 2.0.0   | [PR#468](https://github.com/bbc/psammead/pull/468) Rewrite of inputProvider |